chore,docs: move thai2rom ommx model to anohter file and write docs
